requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.7
Pillow==10.1.0
piexif==1.1.3
python-magic==0.4.27
hachoir==3.2.0
imagehash==4.3.1
python-whois==0.8.0
dnspython==2.4.2
email-validator==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
networkx==3.2.1
matplotlib==3.8.2
PyOD==1.1.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
WTForms==3.1.1
itsdangerous==2.1.2